requirements.txt
appdirs==1.4.3
EasyProcess==0.2.3
enum34==1.1.6
olefile==0.44
packaging==16.8
Pillow==4.1.1
pynput==1.3.2
pyparsing==2.2.0
pyscreenshot==0.4.2
pytesseract==0.1.6
six==1.10.0
virtualenv==15.1.0
